+++ R/jpclust.R
@@ -21,7 +21,7 @@
 #'
 #' Fast C++ implementation of the Jarvis-Patrick clustering which first builds
 #' a shared nearest neighbor graph (k nearest neighbor sparsification) and then
-#' places two points in the same cluster if they are in each other's nearest
+#' places two points in the same cluster if they are in each others nearest
 #' neighbor list and they share at least kt nearest neighbors.
 #'
 #' Following the original paper, the shared nearest neighbor list is