+++ NEWS.md
@@ -6,7 +6,8 @@
 ## Changes
 * Reorganized the C++ source code.
 * README now used bibtex.
-* Removed second sort in kNNdistplot (reported by Natasza Szczypien)
+* Removed redundant sort in kNNdistplot (reported by Natasza Szczypien).
+* Refactor: use more performant anyNA(x) instead of any(is.na(x)) (by m-muecke)
 
 # dbscan 1.1-12 (2023-11-28)